Lift the agent and skill save and autosave handlers into hooks
