+++ packages/twenty-front/src/pages/settings/ai/components/SettingsAgentFormContent.tsx
@@ -1,9 +1,6 @@
-import { CombinedGraphQLErrors } from '@apollo/client/errors';
-import { useMutation } from '@apollo/client/react';
 import { styled } from '@linaria/react';
 import { t } from '@lingui/core/macro';
 import { useContext, useState } from 'react';
-import { useDebouncedCallback } from 'use-debounce';
 import { SettingsPath } from 'twenty-shared/types';
 import { isDefined } from 'twenty-shared/utils';
 import {
@@ -21,30 +18,23 @@ import { SettingsPageContainer } from '@/settings/components/SettingsPageContain
 import { SettingsPageLayout } from '@/settings/components/layout/SettingsPageLayout';
 import { SettingsTabBar } from '@/settings/components/layout/SettingsTabBar';
 import { SettingsRolesQueryEffect } from '@/settings/roles/components/SettingsRolesQueryEffect';
-import { useSaveDraftRoleToDB } from '@/settings/roles/role/hooks/useSaveDraftRoleToDB';
 import { settingsDraftRoleFamilyState } from '@/settings/roles/states/settingsDraftRoleFamilyState';
 import { settingsPersistedRoleFamilyState } from '@/settings/roles/states/settingsPersistedRoleFamilyState';
 import { currentWorkspaceState } from '@/auth/states/currentWorkspaceState';
-import { useSnackBar } from '@/ui/feedback/snack-bar-manager/hooks/useSnackBar';
 import { activeTabIdComponentState } from '@/ui/layout/tab-list/states/activeTabIdComponentState';
 import { useAtomComponentStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomComponentStateValue';
 import { useAtomFamilyStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomFamilyStateValue';
 import { useAtomStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomStateValue';
 import { useSetAtomFamilyState } from '@/ui/utilities/state/jotai/hooks/useSetAtomFamilyState';
-import {
-  type CreateAgentInput,
-  CreateOneAgentDocument,
-  type FindOneAgentQuery,
-  UpdateOneAgentDocument,
-} from '~/generated-metadata/graphql';
+import { type FindOneAgentQuery } from '~/generated-metadata/graphql';
 import { useNavigateSettings } from '~/hooks/useNavigateSettings';
 import { SettingsAgentEvalsTab } from '~/pages/settings/ai/components/SettingsAgentEvalsTab';
 import { SettingsAgentLogsTab } from '~/pages/settings/ai/components/SettingsAgentLogsTab';
 import { SettingsAgentRoleTab } from '~/pages/settings/ai/components/SettingsAgentRoleTab';
 import { SettingsAgentSettingsTab } from '~/pages/settings/ai/components/SettingsAgentSettingsTab';
 import { SETTINGS_AGENT_DETAIL_TABS } from '~/pages/settings/ai/constants/SettingsAgentDetailTabs';
-import { useAutoSaveOnChange } from '~/pages/settings/ai/hooks/useAutoSaveOnChange';
 import { useSettingsAgentFormState } from '~/pages/settings/ai/hooks/useSettingsAgentFormState';
+import { useSettingsAgentSave } from '~/pages/settings/ai/hooks/useSettingsAgentSave';
 import { getSettingsAgentInitialFormValues } from '~/pages/settings/ai/utils/getSettingsAgentInitialFormValues';
 import { getSettingsAiBreadcrumbLinks } from '~/pages/settings/ai/utils/getSettingsAiBreadcrumbLinks';
 import { isDeeplyEqual } from '~/utils/isDeeplyEqual';
@@ -67,11 +57,9 @@ export const SettingsAgentFormContent = ({
   const { theme } = useContext(ThemeContext);
   const { getIcon } = useIcons();
   const navigate = useNavigateSettings();
-  const { enqueueErrorSnackBar } = useSnackBar();
   const currentWorkspace = useAtomStateValue(currentWorkspaceState);
 
   const isCreateMode = !isDefined(agent);
-  const isEditMode = isDefined(agent);
   // Agents created in the workspace belong to its own application; only an
   // installed application's agents are managed elsewhere.
   const isReadonlyMode =
@@ -82,8 +70,6 @@ export const SettingsAgentFormContent = ({
   const [initialFormValues] = useState(() =>
     getSettingsAgentInitialFormValues(agent),
   );
-  const [originalFormValues, setOriginalFormValues] =
-    useState(initialFormValues);
 
   const tabListComponentId = `${SETTINGS_AGENT_DETAIL_TABS.COMPONENT_INSTANCE_ID}-${agentId}`;
   const activeTabId = useAtomComponentStateValue(
@@ -91,16 +77,8 @@ export const SettingsAgentFormContent = ({
     tabListComponentId,
   );
 
-  const {
-    formValues,
-    isSubmitting,
-    handleFieldChange,
-    setIsSubmitting,
-    validateForm,
-  } = useSettingsAgentFormState(initialFormValues);
-
-  const [createAgent] = useMutation(CreateOneAgentDocument);
-  const [updateAgent] = useMutation(UpdateOneAgentDocument);
+  const { formValues, handleFieldChange, validateForm } =
+    useSettingsAgentFormState(initialFormValues);
 
   const settingsDraftRole = useAtomFamilyStateValue(
     settingsDraftRoleFamilyState,
@@ -115,102 +93,17 @@ export const SettingsAgentFormContent = ({
     formValues.role || '',
   );
 
-  const { saveDraftRoleToDB } = useSaveDraftRoleToDB({
-    roleId: formValues.role || '',
-    isCreateMode: false,
-  });
-
   const isRoleDirty =
     isDefined(formValues.role) &&
     !isDeeplyEqual(settingsDraftRole, settingsPersistedRole);
 
-  const saveDraftRole = async (): Promise<boolean> => {
-    try {
-      await saveDraftRoleToDB();
-
-      return true;
-    } catch (error) {
-      if (CombinedGraphQLErrors.is(error)) {
-        enqueueErrorSnackBar({
-          apolloError: error,
-        });
-      } else {
-        const errorMessage =
-          error instanceof Error ? error.message : String(error);
-        enqueueErrorSnackBar({
-          message: t`Failed to save role permissions: ${errorMessage}`,
-        });
-      }
-
-      return false;
-    }
-  };
-
-  const buildUpdateInput = (agentIdToUpdate: string) => ({
-    id: agentIdToUpdate,
-    name: formValues.name || '',
-    label: formValues.label,
-    description: formValues.description,
-    icon: formValues.icon,
-    modelId: formValues.modelId,
-    roleId: formValues.role,
-    prompt: formValues.prompt,
-    modelConfiguration: formValues.modelConfiguration,
-    responseFormat: formValues.responseFormat,
-    evaluationInputs: formValues.evaluationInputs,
-  });
-
-  const autoSave = useDebouncedCallback(async () => {
-    if (
-      !isDefined(agent) ||
-      isReadonlyMode ||
-      !validateForm() ||
-      isSubmitting
-    ) {
-      return;
-    }
-
-    const hasChanges = !isDeeplyEqual(formValues, originalFormValues);
-
-    if (!hasChanges && !isRoleDirty) {
-      return;
-    }
-
-    setIsSubmitting(true);
-
-    try {
-      if (isRoleDirty && isDefined(formValues.role)) {
-        const isRoleSaved = await saveDraftRole();
-
-        if (!isRoleSaved) {
-          setIsSubmitting(false);
-          return;
-        }
-      }
-
-      await updateAgent({
-        variables: { input: buildUpdateInput(agent.id) },
-      });
-
-      setOriginalFormValues({ ...formValues });
-    } catch (error) {
-      enqueueErrorSnackBar({
-        apolloError: CombinedGraphQLErrors.is(error) ? error : undefined,
-      });
-    } finally {
-      setIsSubmitting(false);
-    }
-  }, 1_000);
-
-  useAutoSaveOnChange({
-    autoSave,
-    isEnabled: isEditMode,
-    watchedValue: formValues,
-  });
-  useAutoSaveOnChange({
-    autoSave,
-    isEnabled: isEditMode,
-    watchedValue: isRoleDirty,
+  const { handleSave, isSubmitting } = useSettingsAgentSave({
+    agent,
+    formValues,
+    initialFormValues,
+    isReadonlyMode,
+    isRoleDirty,
+    validateForm,
   });
 
   const canSave = !isReadonlyMode && validateForm() && !isSubmitting;
@@ -237,58 +130,6 @@ export const SettingsAgentFormContent = ({
       Icon: IconTerminal,
     },
   ];
-
-  const handleSave = async () => {
-    if (isReadonlyMode || !validateForm()) {
-      return;
-    }
-
-    setIsSubmitting(true);
-
-    try {
-      if (isRoleDirty && isDefined(formValues.role)) {
-        const isRoleSaved = await saveDraftRole();
-
-        if (!isRoleSaved) {
-          setIsSubmitting(false);
-          return;
-        }
-      }
-
-      if (!isDefined(agent)) {
-        const input: CreateAgentInput = {
-          name: formValues.name,
-          label: formValues.label,
-          description: formValues.description,
-          icon: formValues.icon,
-          modelId: formValues.modelId,
-          roleId: formValues.role,
-          prompt: formValues.prompt,
-          modelConfiguration: formValues.modelConfiguration,
-          responseFormat: formValues.responseFormat,
-          evaluationInputs: formValues.evaluationInputs,
-        };
-
-        await createAgent({
-          variables: { input },
-        });
-        navigate(SettingsPath.AI);
-        return;
-      }
-
-      await updateAgent({
-        variables: { input: buildUpdateInput(agent.id) },
-      });
-
-      navigate(SettingsPath.AI);
-    } catch (error) {
-      enqueueErrorSnackBar({
-        apolloError: CombinedGraphQLErrors.is(error) ? error : undefined,
-      });
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
 
   const handleCancel = () => {
     if (isRoleDirty && isDefined(settingsPersistedRole)) {
